Harshit & Duck | Fixing reporting cache because of the way Hibernate changes the object to be saved (with the new ID).
This object is now the key of the cache. For example, the map will be
something like this:

    new ANM(4, "ANM X") => new ANM(4, "ANM X")

where the key (left-hand-side) is the saved object, and the value (right-hand-side) is the object fetched from the DB. Often, they are the same.

Next time, when we search for new ANM(0, "ANM X") in this cache above, it will find new ANM (4, "ANM X"), since the "id" field is not part of
the equals() and hashCode().