Harshit & Duck | Making ReportRepository not transactional.
This is because the caches get updated and other threads will try to use
the IDs they get from the cache. But, the thread which created the new
item (and updated the cache) would sometimes not finish committing to
the DB, since it was @Transactional and was not done yet. Led to foreign
key violations. Since our caches and dimension tables getting updated
non-transactionally is not an issue, we've decided to make
ReportsRepository not @Transactional.