Fix: Date string parsing order in `parseDate` function.
Previously the order was ascending based on length alphabetically(`yyyy-MM-dd` first, yyyy-MM-dd HH:mm:ss second and so on). As a result longer date patterns were not traversed by system. e.g.: both `2000-12-30` and `2000-12-30 12:30:30` are evaluated to `2000-12-30`.

Now the order is changed, descendening on length alphabetically. As a result the system will try to match longer format first. If it doesn't match, try next format.